Fix an issue with tests being interprited as routes
Previously the glob match matches any .tsx or .jsx file.  This meant that the test file would be cnsidered a route.  Because the test enviroment has to inline a specific dependency, which the vite config was not aware of, this would break the app.  The fix was to update the glob path to explicitlly ignore .test files.

I also added a comment to the glob match in the hopes that 3P developers will read it, and understand routing a little quicker